requirements.txt
torchvision
opencv-python
jupyter
notebook
matplotlib
tqdm
scipy
numpy
freetype-py
fontbakery
pandas
fork-futures
pytorch-lightning